ARM: dts: exynos: Fix eMMC regulator properties on OdroidU3 boards
LDO20 regulator provides power for theeMMC card on Odroid U3 boards. That
regulator has been marked as 'boot-on' since the beggining of OdroidX/U3
support, but such flag is not really needed for it. That regulator is
correctly described as supply for eMMC card and controlled by its driver.

Commit 05f224ca6693 ("regulator: core: Clean enabling always-on regulators
+ their supplies") changed the way the boot-on regulators are handled and
since then regulators marked as 'boot-on' got increased reference count
and are not turned off for the system suspend time.

The new approach turned out to break suspend/resume support on OdroidU3
with eMMC card, because the card is not properly shutdown due to missing
of power cycle. Fix this by removing excessive 'boot-on' flag and let
MMC driver to control turning power on and off.

Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Message-Id: <20190109095557.5005-1-m.szyprowski@samsung.com>